Add files via upload
The code was originally written on Google Colab. The notebook is well documented and contains several markdown cells along with code cells. The dataset has been imported from Kaggle and its link is present in the notebook. The dataset used in the code can be accessed using Kaggle key. Alternatively, the dataset can be downloaded and can be used, in that case code should be modified suitably for it to run error free.